feat: Grammar of Graphics architecture, comprehensive audit, root doc cleanup
- Add Grammar of Graphics spec (composable type-safe visualization pipeline)
- Add Universal Visualization Pipeline spec (barraCuda integration, modality compilers)
- Add Tufte Constraint System spec (machine-checked visualization quality)
- Correct root docs to reflect actual quality metrics (54% coverage, 76 clippy errors)
- Move stale shell scripts to archive, remove empty debris directories
- Fix CHANGELOG stale links to non-existent STATUS.md and NAVIGATION.md
- All 1,511 tests passing, cargo check clean, cargo fmt clean

Made-with: Cursor

+++ crates/petal-tongue-core/src/metrics.rs
@@ -112,11 +112,12 @@ impl CpuHistory {
 
     /// Get the average over all recorded values
     #[must_use]
+    #[allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
     pub fn average(&self) -> f32 {
         if self.values.is_empty() {
             0.0
         } else {
-            self.values.iter().sum::<f32>() / self.values.len() as f32
+            (f64::from(self.values.iter().sum::<f32>()) / self.values.len() as f64) as f32
         }
     }
